NIFI-9625 Refactored Distributed Cache Server and Client Tests
- Replaced TestServerAndClient with separate classes for Set Server and Map Server
- Implemented before and after annotations for starting and stopping server instances
